requirements.txt
colored==1.4.2
Gooey==1.0.8
numpy==1.20.1
pandas==1.2.2
Pillow==8.1.0
psutil==5.8.0
pygtrie==2.4.2
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
wxPython==4.1.0
